requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-bcrypt==1.0.1
requests==2.32.3
python-dotenv==1.0.1
gunicorn==22.0.0
